added a sign up and log in page which is linked with firebase user authentication, & added a sign out button
