+++ .firebaserc
@@ -1,0 +1,5 @@
+{
+  "projects": {
+    "default": "money-wise-32641"
+  }
+}
