+++ pages/index.html
@@ -3,15 +3,19 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Money Wise</title>
    
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="/styles/styles.css">
 <body>
 
     <div id="masterDiv">
 
+        <div id="realHeader">
+            <button id="signOutButton">Sign Out</button>
+        </div>
+
         <div id="titleDiv">
-            <img id="logo" src="logo1.png" alt="Money Wise">
+            <img id="logo" src="/images/logo1.png" alt="Money Wise">
         </div>
 
         <div id="headerDiv"> 
@@ -95,11 +99,7 @@
                 </div>
 
                 
-                <div id="inputsDiv">
-
- 
-
-                </div>
+                <div id="inputsDiv"></div>
 
                 
 
@@ -114,6 +114,6 @@
         
     </div>
     
-    <script src="./script.js"> </script>
+    <script type="module" src="../scripts/script.js"></script>
 </body>
 </html>
