+++ pages/signup.html
@@ -1,0 +1,66 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Sign Up</title>
+    <link rel="stylesheet" href="/styles/signup.css">
+   
+</head>
+<body>
+
+    <div id="masterDiv">
+
+        <div id="titleDiv">
+            <img id="logo" src="/images/logo1.png" alt="Money Wise">
+
+        </div>
+
+        <div id="signupDiv">
+
+            <form id="signUpForm">
+
+                <div  id="signupTitleDiv">
+                    <h1 id="SignUp">Sign Up</h1>
+                </div>
+                
+                <div class="standardDiv" id="entries">
+
+
+                    <div class="inputDiv">
+                        <label for="emailInput" class="labels" id="emailInputText" >Email </label><br>
+                        <input type="email" class="inputs" id="emailInput" name="emailInput" required><br>
+                    </div>
+                    
+
+                 
+
+                    <div class="inputDiv">
+                        <label for="confirmPasswordInput" class="labels" id="confirmPasswordInputText"  >Password </label><br>
+                        <input type="password" class="inputs" id="passwordInput" name="confirmPasswordInput" required ><br>
+                    </div>
+             
+                   
+                </div>
+                
+                <div id="endButtonDiv">
+                    <button id="signupButton2" type="submit">Sign Up</button>
+                    <h3>or</h3>
+
+                  
+                    <button id="loginButton2" type="button">Log In</button>
+                   
+                </div>
+                
+
+            </form>
+
+        </div>
+
+
+    </div>
+
+    <script type="module" src="/scripts/signup.js"></script>
+    
+</body>
+</html>
