+++ styles/signup.css
@@ -1,0 +1,163 @@
+
+#masterDiv{
+    
+    display: flex;
+    flex-direction: column;
+ 
+    align-items: center;
+    background-color: #f3f5f6;
+    width: 100%;
+    height: 100%;
+    position: relative;
+}
+
+body, html{
+    height: 100%;
+    width: 100%;
+}
+
+#titleDiv{
+
+
+    width: 90%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    
+}
+
+#logo{
+
+    width: 100%;
+    max-width: 2000px;
+
+}
+
+#signupDiv{
+    background-color: #e6e3e0;
+    width: 700px;
+    height: 595px;
+    border-radius: 10px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    box-shadow: 5px 5px 15px #888888;
+    margin-top: 100px;
+ 
+}
+
+#signupTitleDiv{
+
+    display: flex;
+    flex-direction: column;
+    align-items: center;;
+}
+
+.standardDiv{
+    display: flex;
+    flex-direction: column;
+    background-color: #e6e3e0;
+    position: relative;
+    align-items: center;
+   
+    width: 100%;
+    position: relative;
+   
+    
+}
+
+#SignUp{
+
+    font-size: 75px;
+    color:rgb(69, 65, 65);
+    margin-top: 20px;
+
+
+}
+
+
+.inputs{
+
+    width: calc(100% - 10px);
+    height: 40px;
+    border-radius: 8px;
+    border: none;
+    font-size: 20px;
+    margin-top: 5px;
+    margin-bottom: 30px;
+    padding-left: 10px;
+}
+
+.inputDiv{
+    width: 85%;
+}
+
+
+#signUpForm{
+   
+    width: 100%;
+}
+
+.labels{
+    font-size: 30px;
+    color:rgb(69, 65, 65);
+  
+}
+
+#endButtonDiv{
+
+    margin-top: 15px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+}
+
+#loginButton2{
+       background-color: rgb(69, 65, 65);
+    color: white;
+    font-size: 22px;
+    width: 20%;
+    height: 50px;
+    font-weight: 700;
+    border-radius: 8px;
+    transition: all 0.1s;
+    
+}
+
+#signupButton2{
+
+    background-color: rgb(69, 65, 65);
+    color: white;
+    font-size: 22px;
+    width: 85%;
+    height: 50px;
+    font-weight: 700;
+    border-radius: 8px;
+    transition: all 0.1s;
+}
+
+#loginButton2:hover{
+
+    background-color: white;
+    border: none;
+    color: rgb(69, 65, 65);
+}
+
+#signupButton2:hover{
+    background-color: white;
+    border: none;
+    color: rgb(69, 65, 65);
+}
+
+h3{
+    color: rgb(69, 65, 65);
+    font-size: 25px;
+}
+
+
+
+
+
+
